JCBC-1559: Cache Bucket, Collection and Scopes
This changeset adds - at each level and for each blocking, async
and reactive - a cache that caches the underlying instance. While
we recommend users to reuse the instance, people might not do it
all the time and to help them not create unnecessary garbage we
can cache the instances before handing them out.

They are stored in concurrent hash maps and computed lazily, so
the overhead should be minimal even if accessed a lot.

Change-Id: Ifc9fc4a7e43d92ef7b7106b3cd2c558bbf45985f
Reviewed-on: http://review.couchbase.org/120095
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>