JVMCBC-795: Validate service not available with current config
Motivation
----------
Changeset #5030622 introduced the service not available exception
if no node has the service enabled the request is looking for. This
works fine during steady state but has problems during bootstrap:

There exists a time delta where we have a "good" config which shows
that the service is enabled but the actual node is still being
bootstrapped into and might not reflect that config.

Modifications
-------------
In order to handle the situation described above, instead of killing
the request immediately we double check with the current configuration
if the service really is not enabled in neither the global nor the bucket
configurations. Only if it cannot be found there too we know that
the service really does not exist in the cluster to the best of the
sdks knowledge.

Note that this bug was actually cought with log output in one of the
integration tests:

21:42:35 WARN  [com.couchbase.config:513] [com.couchbase.config][BucketConfigRefreshFailedEvent]
Reason: STREAM_FAILED, Type: MANAGER, Cause: com.couchbase.client.core.error.ServiceNotAvailableException:
The mgmt service is not available in the cluster. {"completed":true,"coreId":8,"idempotent":true,"requestId":42,
"requestType":"BucketConfigStreamingRequest","retried":0,"timeoutMs":75000} {"coreId":8}

Which shows the issue since the management service is required to be there
on every node but might not be bootstrapped yet.

Change-Id: I75b0c0fb11f26ba8faee7e1db650e3cecf74f38c
Reviewed-on: http://review.couchbase.org/120180
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>