JVMCBC-799: Increase netty leak detector sample interval
The sample interval by default is 128, which triggers the detector
quite often under higher load. This changeset increases the interval
if not set by the user (so it can still be overridden) so that not
that many exceptions are generated (and because we are not carrying
around pooled buffers as much as before the likelyhood if a leak is
also greatly reduced).

Note that we are also planning to disable it completely later on
if we do not find any leaks reported in the field with this new
SDK.

Change-Id: I91178f1b21f8c0e14f1238a7116c49125d08b645
Reviewed-on: http://review.couchbase.org/120200
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>