JVMCBC-802: Do not retry view request when design doc is found but view is not
When porting the retry logic from SDK 2 for views it hasn't been ported
100% correctly. SDK2 checks that the reason is only "missing", not that
it just contains the word. In the case mentioned in the title, the word
missing also shows up leading to false retries of nonexistent views.

Relevant SDK2 code:

content.contains("\"reason\":\"missing\"")

Note that we split up the reason in its own field now, so we need to
exact match on the "missing" word in it.

Change-Id: I384f970b7abd1e07237800071e8bbf8dace887c1
Reviewed-on: http://review.couchbase.org/120211
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Build Bot <build@couchbase.com>