JVMCBC-803: Use connectTimeout for config loading
Instead of relying on the kv and management timeout, make sure that
the loader timeouts are connected to the connecting timeout instead.
This makes sure that even if a user tunes the kv timeout low we are
still a little more reliable on the internal side.

Note that using the same connect timeout as in the endpoint uncovered
a race condition that led to a harmless but ugly onErrorDropped
in reactor. So this change also performs the "unsubscribe" check
with the mono so that if the higher level timeout unsubscribes
the future the result is ignored.

Change-Id: I71119dacf1237d39e05c20232e675f74cc730afb
Reviewed-on: http://review.couchbase.org/120238
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>