SCBC-158: Handle 'too many set inserts' error
jsoniter-scala is a library used internally when converting JSON
to/from a user's case class.

If the case class contains a Scala Set or Map, there's a 1,024 limit on
the number of items that can be in the collection.

I'm adjusting that arbritrary default to a higher arbritrary default of
100k.  Have also asked the developer if there are any downsides to this,
and will return to this decision if needed.

Unfortunately the default has to be duplicated in 3 places, as attempts
to factor it out cleanly results in a strange macro compilation error.

Change-Id: I60642e97c02456f3982ef54ddb22969ab7989891
Reviewed-on: http://review.couchbase.org/120458
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>