JVMCBC-808: Trim down netty
This changeset modifies the dependencies to not include netty-all,
and instead only what we really need. Http (which has the rest
transitively included) and both the native epoll and kqueue
transports.

This trims down the JAR by 1MB, many more MBs uncompressed. It also
just in general helps reducing shipping code we do not need.

Change-Id: Idf52a9b0df447f6c16c3f1b3e014cfea6204bf8f
Reviewed-on: http://review.couchbase.org/120849
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Build Bot <build@couchbase.com>