JCBC-1574 JsonObject/Array.put(String,Object) fails on Map or List
Motivation
----------
There already exist specialized methods for putting Map and List values,
but they are not invoked if the exact type of the value is not known
at compile time.

Support passing a Map or List to JsonObject/Array.put(String,Object).
This is important since these methods are used by Spring Data Couchbase.

Modifications
-------------
Add a JsonValue.coerce(Object) method that converts the given object
to a type that can be put in a JsonArray or JsonObject. Use this method
everywhere a value of unknown type is added.

For consistency, convert any ClassCastExceptions to
InvalidArgumentExceptions.

Change-Id: I31fe99780ab5a3c402e427d8f0d4e47f9d5558b5
Reviewed-on: http://review.couchbase.org/120912
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>