JCBC-1578: Bring back Transcoder for Java Object Serialization
This changeset brings back the functionality of SDK 2 which allows
you do work with objects directly that support java object serialization.

This is distinct from the JSON ones in that it supports any object that
implements Serializable, but it also means that the result value will
be stored as a binary blob on the server. Effectively it can be (and is)
used for caching scenarios (i.e. the couchbase spring cache implementation).

Change-Id: I4988d01d3e2ba54421da8ef9525de1c9f4180b19
Reviewed-on: http://review.couchbase.org/121142
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>
(cherry picked from commit 6aef0e3dbdbc6e4aaae009920ebff50cfe5a37fa)
Reviewed-on: http://review.couchbase.org/121443
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: David Nault <david.nault@couchbase.com>