JCBC-1580: Projections must not import unbundled jackson
This leads to NoClassDefFoundError at runtime if a custom
jackson is not on the class path - making matters worse it
is not even used in this class and looks like leftovers
from earlier iterations.

Note that it has not been caught in testing because the
unbundled jackson is on the classpath since we need
it to compile (optional dependency).

Change-Id: I4969851216175e6349f21b5f48539e2bd947f771
Reviewed-on: http://review.couchbase.org/121320
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>