JCBC-1582: Use client context ID for PREPARE query
This changeset makes sure that 1) the original user context ID
is used in a PREPARE statement and 2) if none is provided generates
a UUID one (like the regular behavior would).

This also implicitly makes sure that when an exception is raised
for the query service it shows up in the error context even when
the issue itself is the prepare statement.

Change-Id: I4b100858039fed5e5b557fad7a9cfa46e2c1ceeb
Reviewed-on: http://review.couchbase.org/121411
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>
Reviewed-on: http://review.couchbase.org/121668
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>