SCBC-201: Expose Java core environment through Scala ClusterEnvironment
The Scala ClusterEnvironment constructs a Java core-io environment under
the hood, which is currently private.

This prevents the user from configuring the event-bus, so exposing the
core directly here.  It would also be possible to only expose the
event-bus, but this is more flexible and I don't see any harm to it.

Change-Id: If0cf4dbb09611fea6fa2b4658408851e091f8aa5
Reviewed-on: http://review.couchbase.org/121566
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>