SCBC-204, JCBC-1585, JVMCBC-812: Add support for new server flag createAsDeleted
As part of the work for TXNJ-125, the server is exposing a new flag,
createAsDeleted, that will allow creating documents in a 'tombstone'
state when using Sub-Document mutateIn.

This changeset exposes this in Java and Scala as @Stability.Internal.

(Scala isn't strictly needed here, but to be consistent.)

Note, this requires a particular version of the cluster.  This is
likely to be 6.5.1+.

Change-Id: Icf6aee594678c6f9b00b1db7652d8ed2331a5e2c
Reviewed-on: http://review.couchbase.org/121618
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>