JVMCBC-819: Distinguish bucket not found with select bucket.
When select bucket fails with a 0x01 error, we can turn that into
the equivalent of a "bucket not found exception" instead of a more
generic one. This will make it more obvious when looking at the logs
to see what's going on.

This error is especially important because we're likely going to see
it bootstrapping against non-kv nodes.

Note that separate changesets are going to look into improving the
non-kv node bootstrap experience, but this is a "quality of life"
change nonetheless.

Change-Id: Iff73fb81e64f9ea15035e9856b1c321692939ace
Reviewed-on: http://review.couchbase.org/121910
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>