JCBC-1587, SCBC-207: Exists must return false when preceded by remove
It turns out the server returns success when GetMeta immediately
follows a remove, but logically for the user it is still deleted.

This changeset parses the deleted flag out of the kv server response,
exposes it on the GetMetaResponse and then makes use of it in both
the java and scala code. Unfortunately the integration tests did not
cover this specific case, so they have been added as regression tests.

Change-Id: Id1213c2fae8c6f53d0809124bfd9d11b597e28a6
Reviewed-on: http://review.couchbase.org/121961
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>