JCBC-1588: Give TypeRef overload of JsonSerializer a default "unsupported" impl
To make it easier for the application to create a custom JsonSerializer,
provide a default implementation for the overload that takes a TypeRef.

The impetus for this change is that trying to write a custom serializer
for MessagePack required some non-trivial marshalling and casting, to
turn the TypeRef into something usable by the MessagePack Java library.

The overload is only needed if the application is using
contentAs(new TypeRef<..>{}) - which is an unusual requirement.

Change-Id: I9c084174a2b18fdde73bfee27f9729e71601dc82
Reviewed-on: http://review.couchbase.org/122259
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>