JCBC-1604: Properly clear cache when repreparing query
Motivation
----------
In internal situational testing we observed that during query
swap rebalance, n1ql queries are not properly re-prepared if
the server returns 4040 and similar. The operation it self
was retried, but it needs to do more internally (clear cache,
re-prepare) to remedy the situation transparently.

Modifications
-------------
This changeset is twofold:

 - In core, it explicitly does not retry the prepared error codes
   that must be handled at the upper level.
 - In the java QueryAccessor, additional retry handling code
   has been added which takes those errors into account and if
   they are raised consults the retry strategy to check if the
   request can be retried. If it can, it delays it and then
   performs the whole re-prepare cycle.

Change-Id: I591602c6649166f40821234a7017dccf403b4cc1
Reviewed-on: http://review.couchbase.org/123130
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>