JVMCBC-826: Fix NPE when authenticating through LDAP
Motivation
----------
When the server has LDAP authentication enabled, over KV it will
only allow PLAIN authentication. Since PLAIN is disabled by default
no set of SASL mechanisms can be determined.

The code right now does not handle this case and throws a NPE.

Modifications
-------------
Instead od failing with an NPE, the code now explicitly fails
the connection with a descriptive error message instructing the
user what to do next (either use TLS which is preferred) or
enable PLAIN on the password authenticator (insecure).

This changeset also takes the opportunity to add javadocs to
the password authenticator and add a convenience overload to
enable the PLAIN mechanism.

Change-Id: Idaefe0baa92b28c09a162ebe87ecf2205d2ba6b2
Reviewed-on: http://review.couchbase.org/123367
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>