JVMCBC-828: Omit internal config request in orphan reporting
This changeset introduces a marker interface called
"UnmonitoredRequest" that if present on a request will make
it invisible to certain logging/tracking subsystems.

Especially these requests are not included in orphan reporting
and if they are cancelled they are only logged at debug, not
info level.

For now only the carrier*config requests have the marker interface
on, because especially during polling in slow systems they will
spam the orphan reporter and also the cancellations in the logs.
We can put additional instrumentation on the refreshers if needed
which will also provide more accurate debugging infos.

Change-Id: Ia523341ea740c4ab4b7b0499d1e0a49073a36b43
Reviewed-on: http://review.couchbase.org/123389
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>