JVMCBC-830: More convenience overloads for certificates
Based on forum feedback it has been identified that the current
overloads are not as convenient, especially when converting from
SDK2 to SDK3.

This changeset adds more overloads to initialize both the
SecurityConfig and the CertificateAuthenticator directly from
a KeyStore/TrustStore and it also improves the javadocs all
around a bit.

Change-Id: Iedf0d3b412b79960ec7ece95a8edd0ed57380252
Reviewed-on: http://review.couchbase.org/123849
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/124718
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>