JVMCBC-834: Retry operation if collection not found
Previously a CollectionNotFoundException would only trigger a retry if
a collection refresh is in progress. But it can also happen that if
a collection has been created in the meantime the collection is
not found but no refresh is in progress.

This changeset modifies the code so that it will always trigger a
retry, and if no collection refresh is currently in progress it will
proactively trigger a new one.

Change-Id: Id442b9f6283630196151b747aa99ce63340db8a1
Reviewed-on: http://review.couchbase.org/124108
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>