JCBC-1617: Fail with explicit FeatureNotAvailable for collections on older clusters.
For all clusters that do not support collections (or do not have it enabled
by default like 6.5) the collection manager must return a proper exception
(FeatureNotAvailable). Before this change it either failed with a generic
exception or with a jackson decoding error depending on the code path.

Note that the CollectionsNotAvailableException has been removed since it
is not used anywhere and we should use the FeatureNotAvailable anyways
since that one is standardized across SDKs.

Change-Id: I54af78a8f8b7748251a3d807ac5151533b65141f
Reviewed-on: http://review.couchbase.org/124323
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>