JVMCBC-835: Fail fast if user connects to 8091 in connstr
Motivation
----------
Over the last couple weeks a pattern emerged many times: userss
would use "localhost:8091" as a connection string with good intentions,
but it would actually set the kv bootstrap port to 8091 which is
of course not what we want.

Modifications
-------------
To prevent this from happening again, the code now checks for this
condition, fails fast and also provides guidance on how the connection
string should look like instead.

Also note that a minor bug introduced with a previous changeset has
been fixed that if no explicit port type is found the KV one is used
implicitly.

Change-Id: Ibef85a0c99653d19a28839579d1c1b346b7dfb6f
Reviewed-on: http://review.couchbase.org/124364
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/124726
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>