Automation: Build the Java client release archive in the package phase
Motivation
----------
Make it easier to generate the Couchbase-Java-Client-<version>.zip
release archive.

Modifications
-------------
Add a 'release-archive' assembly so you can generate the release
archive by running "mvn package" (from either the top level or
the level or the java-client module directory).

Limitation: POMs for core-io and java-client are not included in the
archive.

Change-Id: Ie749857adf6fcb7b6546a041f51f5f04734a2f48
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/125435
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>