JVMCBC-797: Optimize reading flags and extras as single int.
This changeset is a pure (and small) performance optimization guided
by JFR profiling. Running a get workload it can be observed that since
we need to read the extras on every result to extract the flags:

 - Integer autoboxing happens in the closure
 - A slice needs to be created to extract the extras

This can be simplified with a specialized method that can also avoid
both of the "issues" above. You won't notice it in performance numbers
on its own, but it produces a little bit less gc when running thousands
of requests per second.

Change-Id: I293f793e1a310b830a79d0669bf235e20baedfe2
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/125488
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>