JVMCBC-845 - Alternate fast forward and active location.
Motivation
----------
It has been observed that under the following (although considered
edge case) secnario:

 - Rebalance stopped in the middle, so a ff-map stays around for
   a long time and does not move
 - An operation already being retried (> 0 retry attempts)

that the operation might time out and never complete. This is due
to the fact that the current implementation, if retry > 0 and
ff map present, will always choose the ff map.

Modifications
-------------
This changeset slightly modifies the logic that the ff map will
be chosen based on the number of retry attempts only in the "odd"
case, so only in times where the retry attempts is > 0 and it is an
odd number (n % 2 == 1) the ff map will be tried.

This makes sure that even if the map never moves forward and the
operation has been retried it will make progress eventually and
complete.

Change-Id: Ia8acc29cb1cca2baca5a28353b31a489cbf5d97b
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/126015
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>