JVMCBC-846: Deprecate XattrInvalidFlagComboException and XattrInvalidOrderException
This changeset deprecates and effectively removes two exceptions from being thrown.
They should not come up in the first place because the SDK must prohibit those states
in the first place. They are deprecated because the SDK-RFC removed them.

We are not removing them, since users (although unlikely) might have a catch on
them in their codepath.

Change-Id: I87334950c238c0f76adcaa9e9659fc9bc9432f91
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/126017
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>