JCBC-1637 Expose ConfigurationProvider seed nodes as Flux
Motivation
----------
Provide a way to get the addresses of all nodes running
the KV service, with alternate addresses resolved.

This helps connectors bootstrap the DCP client.

Modifications
-------------
Expose the ConfigurationProvider's seed nodes as a Flux.

Since the goal is to get a set of as many seed nodes as possible,
the initial set of bootstrap seed nodes provided by the user
is not published.

Make DefaultConfigurationProvider's current set of seed nodes
unmodifiable so it can't be altered by an event bus listener
or Flux subscriber.

Make SeedNode.LOCALHOST unmodifiable.

Change-Id: Ib42cf34fc692c5d615b90dae9318f8a8590b2461
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/129110
Tested-by: David Nault <david.nault@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>