JVMCBC-693: Pipeline first 3 requests of KV bootstrap.
Motivation
----------
In an effort to speed up kv bootstrap, the SDK 3 bootstrap RFC
mandates that the individual kv bootstrap steps are pipelined as
much as possible.

Modifications
-------------
This changeset begins the refactoring of the kv bootstrap pipeline
and starts with the "low hanging fruit", the first 2 handlers in
the sequence: HELLO negotiation and error map loading.

Before this change, the handlers would only fire channelActive
on the context once a response is received. Now, channelActive
is fired immediately after the handler sends its initial request
on the wire. Since there is no "back and forth" needed for them,
they can be pipelined nicely.

With this change applied, the initial hello, the error map loading
and the next level SASL list mechs are pipelined, so the latency
is reduced to one roundtrip vs. 3.

The whole SASL part as well as the following select bucket will be
pipelined in a later changeset.

Change-Id: I65b8a7aabfa77f7b4252eaed400237c8ae8c1c1e
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/129148
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/129282
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>