JVMCBC-856: Just opened connection in pool must not be cleaned up.
Motivation
----------
In debugging it has been discovered that the current pool cleaner
logic had a bug: it would work under the assumption that if it is
"free", a request must have happened so the lastResponseReceived
timestamp is available.

This is not the case though, since a request is rescheduled while
the endpoint gets opened in parallel, but it can happen that the
rescheduled request makes its way to a completely different endpoint
on the next try, leaving this one hanging around until a request
gets dispatched eventually (or never).

When the idle cleaner comes along every interval (5s), the logic
mentioned above caused "fresh" sockets to be cleaned up as well
if no request had ever been dispatched into it.

NOTE: at this point it is clear that this is a logic bug, but the
actual impact on user workloads is not yet fully determined.
Regardless, this needs to be fixed.

Modifications
-------------
To prevent the issue outlined above, the endpoint itself now also
stores an absolute timestamp when it got last CONNECTED. The cleaner
logic on the service level now makes sure that if no request ever
got dispatched (so lastResponseReceived is 0), it takes this time
as the idle time to check instead of always cleaning it up
prematurely.

Result
------
Fresh sockets in the pool where no request has been dispatched yet
is not cleaned up prematurely.

Change-Id: I4fb5ce1db0b0192251b0e9e675b32fc395d8bde4
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/130529
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>