JVMCBC-860 - Add explicit config for ThresholdRequestTracer
Motivation
----------
Right now the threshold tracer cannot be configured by just
applying a config, similar to how the orphan reporter can
be configured. This makes it very hard to actually customize
it, since it has dependencies on the event bus which in turn
has a dependency on the scheduler.

Modifications
-------------
This changeset adds a specific config for the tracer, but adding
it in a backwards compatible way so you can still configure it
through the builder if needed.

Now it is possible to just provide a config from the environment
and it will be picked up automatically. Also, the config is
logged on startup and can be used with all our custom property
config mechanisms.

Change-Id: I73426f95b78107f0f245b5505ac8d5a809f20741
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/130754
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>