JVMCBC-862: Fix race conditions with identifiers on bootstrap
Motivation
----------
It has been found that when node identifiers switch (i.e. from
a bootstrap hostname to a ip from the server config or if
alternate addresses are used), in-flight targeted request
might never get dispatched since their target node has been
concurrently removed from the node structure (since a config
arrived from a different bucket or a global config, and
removed the matching identifier node).

To address this issue, a couple spots in the code need to
be modified to handle this race condition.

Modifications
-------------
At the core, the locators now look at the request and if it
is a TargetedRequest and it cannot be dispatched, further
investigation is performed: all the nodes are iterated and
the code checks if the identifier is even part of the cluster.

If not, the request is cancelled with a specific reason since
in this case it may never be able to make progress in the
first place.

Since now targeted requests might get cancelled early, the
config provider needs to handle this case and if so retry
the whole loading phase with a new set of seed nodes that
are now fresh. Some of the config provider code had to
be refactored so that retry and resubscribe actually
picks up fresh seed nodes.

Change-Id: I05b059893a4e3454de1010709bd13e428541714b
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/131155
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/131447
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>