JCBC-1645 Expiry durations longer than 30 days are interpreted incorrectly
Motivation
----------
Couchbase Server interprets the "expiry" field different depending
on whether it's less than 30 days. If it's less than 30 days, it's
interpreted as a relative offset in seconds from the current time.
Otherwise it's interpreted as an absolute timestamp in UNIX epoch seconds.

Prior to this change, the SDK would pass the number of seconds in the
duration directly to the server, causing documents to expire
prematurely for durations longer than 30 days.

Modifications
-------------
When generating the request, convert expiry durations longer
than 30 days into absolute timestamps. (Unless they're longer
than 50 years, in which case treat them as absolute for
bug-compatibility with previous 3.0.x versions.)

Recompute the expiry every time a request is executed. This way
the effective expiry duration is relative to the time the request
is executed, not the time the options block was created.

Change-Id: I5d7090e789af45d83a61cd1cdff99d8270d7948a
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/131215
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>