JVMCBC-863: Bucket-level ping report includes other buckets as well
Currently a bucket-level ping will include all buckets for view
and KV services, which is not how it should be. If you ping at
the bucket level only the specific bucket in scope should be used.

This changeset modifies the API and inverses it so that instead
of "clusterLevel" as a boolean it allows to optionally pass in a
bucket name. If none, cluster-level is assumed. This way we can
filter the configs based on the name and as a result only the
bucket in scope is part of the report (and we also potentially
end up pinging less endpoints with it).

Change-Id: If8bb2079c5a3730833820dc856050a9f0e2155bb
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/131263
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/131448
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>