Add possibility to override test properties through env
This changeset makes it possible to override certain test
properties at the CLI. So for example if you want to run the
full test suite against a real cluster and not the mock, now
you can do:

CLUSTER_TYPE=unmanaged mvn clean test

The rules are: it takes the env key, converts it to lowercase
and replaces underscores with dots. So CLUSTER_TYPE becomes
cluster.type.

Change-Id: I96d2de0dabfea35f0fefd5d773c0645c8726b9d5
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/131441
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>