JVMCBC-864: Extract SASL List Mechanisms
This changeset is a preparation step to further pipeline the SASL
stages. The SASL list request/response has been refactored out into
its own handler while still leaving the current one in place. Also,
the new handler is not added to the pipeline yet since it would cause
two SASL list mechs request which is unnecessary.

A later commit will make use of this handler as soon as the current
mechs code has been removed from the primary sasl auth handler.

As a nice bonus, the unit tests are better now for this command in
isolation.

Change-Id: I83fb25b9ebff915658bb517770659b0beb163bde
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/131626
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>