JVMCBC-864: Try SASL auth mechanisms in a best effort fashion
This changeset builds upon the previous list mechs refactoring
and adds the following:

 - It does not wait for the sasl list mechs to complete, but rather
   just starts the SASL auth phase with the list of supported
   mechanisms
 - Since the list could be wrong, on auth failure, new code has been
   added which then goes back and compares the negotiated list with
   the selected mechanism
 - If we find that the mechanism was allowed, we fail the auth since
   then it's a legit auth error
 - If we find that the needed mechanism has not been used, we re-do
   the full auth process with the proper mechanism based on the sasl
   list mechs payload

This gives us a chance to pipeline as much as possible and only
fall-back to a re-auth if the mechanisms are not correct (which
at the moment only happens in a case where LDAP is enabled and no
TLS is used).

Note that while this change pipelines two more ops (the sasl list
mechs and the initial SASL auth step) there are more enhancements
to come w.r.t pipelining of the select bucket step at the end of
the SASL sequence in a later commit.

Change-Id: I1a63990d336439868b61f141165eb787a5dd1df4
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/131627
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>