JVMCBC-864: Pipeline Select Bucket if needed
This is the last changeset in the KV pipeline series which enables
pipelining of the last request (select bucket) if happens to
be in the pipeline (which it is for all bucket-level kv connections).

The SASL auth handler now tracks the number of roundtrips that need
to be performed for the active SASL mechanism and if there is only
one roundtrip left it sends the channel active up the pipeline to
trigger further pipelining with higher level handlers (i.e. the
select bucket request will be pipelined right after the last challenge-
response request). If a reauth is needed, the channel active will
be called multiple times, triggering multiple select bucket
calls which is ok.

Note that because select bucket is now pipelined, the feature
negotiation is not yet complete, so I had to get rid of the
code which checks if select bucket is negotiated. It is not
needed anyway since all of the cluster versions we support
have it (and it is a mandatory requirement for RBAC).

Change-Id: I310cc4268eb1c327d17b13d5171f6f72b8dd6e66
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/131628
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/132018
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>