JCBC-1661 Discourage using GetResult.expiry() and add alternative
Motivation
----------
The Duration returned by GetResult.expiry() is bogus -- it represents
the time span between the start of the epoch and the instant when the
document expires.

Changing the return type of the existing expiry method to Instant
would be a breaking change.

Modifications
-------------
Add "uncommitted" GetResult.expiryTime() which returns an Instant.

Discourage using GetResult.expiry() and update the docs
to indicate what it's really doing.

Add some Javadoc explaining these methods always returns
an empty Optional unless GetOptions.withExpiry(true) is used.

Change-Id: I52a51baf7864f30e23881387c8131102499b1b0d
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/132051
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>