JVMCBC-867: Do not grab ByteBuf slice when parsing server latency
During profiling with JFR I found that most of the allocations
(when doing gets in a tight loop) on one of the IO threads is
actually allocating a slice from the buffer pool for each
server duration decoding. Since all we need is an offset into
the response we can do the checking without allocating a slice.

This reduces the allocations in a 6 minute run by 30% in the
cb-io-kv-* thread where the op is performed. Your mileage will
vary but overall this is another small step to improve
efficiency.

Change-Id: I1601297eb4604e77a1f3c6ef9baf277e94e2648e
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/132305
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>