JCBC-1665 Allow specifying absolute expiry time
Motivation
----------
For some use cases is might be more appropriate to specify the expiry
as an Instant.

Support round-tripping of expiry time using Instant value GetResult.

Modifications
-------------
Add an Expiry class that can be either absolute or relative
and knows how to encode itself to the kind of value
expected by the server.

Modify options classes to represent expiry as Expiry
instead of Duration.

Remove ExpiryUtils, since that logic is now in Expiry.

Change-Id: Ifb54ef9cc24b2abed07fe8ce35b8a33a6df47c31
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/132342
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>