Gardening - getting CreateAsDeletedIntegrationTest passing again
There seems to have been a minor server change here.  A test
that was previously passing now returns a "document is not
json" error.

However, it doesn't matter.  It only happens when the
createAsDeleted flag (new for 6.6) is used, and only when
xattrs are not used.  The only user of createAsDeleted is
expected to be transactions, which exclusively will use
xattrs.  So, just modifying these tests after reporting the
issue to memcached team.

Change-Id: I65771f91cadd928ed39b949e2287a0dcab3e532b
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/133039
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>