JCBC-1668: Fix NPE when toString on SearchRow
This changeset fixes an NPE when toString or fieldsAs is called
when no fields are present in the result. In this case just null
should be returned instead of a NPE deep doen in the jackson
serializer stack.

As a drive-by fix I noticed that the equals and hashCode are
old because they do not take all fields into account, so
regenerated them as well.

Change-Id: I9f150ced478a4a0873ebb003166f858d1f091d0e
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/133142
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>