JCBC-1670 Add WaitUntilReadyOptions.serviceTypes() overload for varargs
Motivation
----------
In many places in the API, methods that accept a Set have overloads
that accept the elements as varargs.

As a convenience for users, do the same thing for
WaitUntilReadyOptions.serviceTypes()

Modifications
-------------
Add WaitUntilReadyOptions.serviceTypes(ServiceType...)
which delegates to serviceTypes(Set<ServiceType>).

Instead of throwing IllegalArgument for empty set,
interpret it as "wait for all possible services".

Change-Id: I81dbdb30424983e0e88143a680fdf587d9e02f06
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/133178
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>