JCBC-1678: Remove deprecated use of reactor retryWhen
Reactor 3.4.0 will remove the retryWhen overload with a function
which is heavily used inside the SDK. This changeset adds a
utility to the Retry helper class which allows to turn it into
the new reactor retry.

The correct fix would be to change all our uses to the reactor
retry but it is not a simple change-and-replace. Since we are
late in the release process I've been aiming for minimal
impact.

Another change can then be made on top to convert it over and
ultimately deprecate our own Retry class we ported over.

Change-Id: I49c59e4e03989139f2b85b4e3cbce578b921095c
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/133610
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Build Bot <build@couchbase.com>